// predicate-builder-service/src/enums/operator.ts
export enum Operator{
    EQUALS = "EQUALS",
    CONTAINS = "CONTAINS",
    ISNULL = "ISNULL",
    STARTSWITH = "STARTSWITH",
    ENDSWITH = "ENDSWITH",
    GREATERTHAN = "GREATERTHAN",
    GREATERTHANANDEQUALS = "GREATERTHANANDEQUALS",
    LESSTHAN = "LESSTHAN",
    LESSTHANANDEQUALS = "LESSTHANANDEQUALS",
}